graphql auto-registration; enhance response object mapping

+++ user-service/src/main/java/com/listing/user/service/repository/UsersEntityRepository.java
@@ -4,6 +4,6 @@ import com.listing.user.service.entity.UsersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
-public interface UsersEntityRepository extends JpaRepository<UsersEntity, Long> {
+public interface UsersEntityRepository extends JpaRepository<UsersEntity, Integer> {
     Optional<UsersEntity> findUsersEntitiesByEmail(String email);
 }
