+++ user-service/src/main/java/com/listing/user/service/service/DatabaseService.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public interface DatabaseService {
     UsersEntity createUser(UsersRequest usersRequest);
     List<UsersEntity> findAll();
-    Optional<UsersEntity> findById(Long id);
+    Optional<UsersEntity> findById(int id);
     Optional<UsersEntity> findUserByEmail(String email);
     EntityTypesEntity createEntityType(EntityTypesEntity entityTypesEntity);
     List<EntityTypesEntity> listEntityTypes();
