+++ user-service/src/main/java/com/listing/user/service/service/UsersEntityService.java
@@ -12,6 +12,6 @@ public interface UsersEntityService {
 
     UsersEntity createUser(UsersRequest usersRequest);
     List<UsersEntity> findAll();
-    Optional<UsersEntity> findById(Long id);
+    Optional<UsersEntity> findById(int id);
     Optional<UsersEntity> findUserByEmail(String email);
 }
